cleaned up inference.py, revamped config file
