+++ model/README.md
@@ -1,4 +1,10 @@
 # PokeGenerator Model
 
+## Configuration
+```vim config.json```
+
 ## Training
 ```python3 train.py```
+
+## Inference
+```python3 inference.py```
