--- model/training_config.json
@@ -1,8 +1,0 @@
-{
-  "batch_size": 32,
-  "checkpoint_directory": "checkpoints",
-  "dataset_cache": "/datasets/Pokemon_Sprites/dataset_128_cache.npy",
-  "dataset_directory": "/datasets/Pokemon_Sprites/dataset_128",
-  "epochs": 500,
-  "validation_split": 0.1
-}
